requirements.txt
fastapi[all]
SQLAlchemy
uvicorn
psycopg2
python-jose
bcrypt
passlib